Initial structure change for project.
