--- .env.sample
@@ -1,3 +1,0 @@
-REACT_APP_CLIENT_ID=example
-REACT_APP_OKTA_ISSUER_URI="https://example-882474.okta.com"
-REACT_APP_API_URI=http://localhost:8005
