+++ src/components/exampleDataViz/components/index.js
@@ -1,0 +1,3 @@
+import DataVizContainer from '../pages/ExampleDataVizContainer';
+
+export { DataVizContainer as ExampleDataViz };
