+++ src/components/exampleList/components/index.js
@@ -1,0 +1,1 @@
+export { default as ExampleListPage } from '../pages/ExampleListContainer';
