+++ src/components/home/components/RenderHomePage.js
