+++ src/components/home/components/index.js
@@ -1,0 +1,1 @@
+export { default as HomePage } from '../pages/HomeContainer';
