+++ src/components/home/pages/HomeContainer.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useMemo } from 'react';
 import { useOktaAuth } from '@okta/okta-react';
 
-import RenderHomePage from './RenderHomePage';
+import RenderHomePage from '../components/RenderHomePage';
 
 function HomeContainer({ LoadingComponent }) {
   const { authState, authService } = useOktaAuth();
