+++ src/components/landing/components/index.js
@@ -1,0 +1,1 @@
+export { default as LandingPage } from '../pages/LandingContainer';
