+++ src/components/landing/pages/LandingContainer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import RenderLandingPage from './RenderLandingPage';
+import RenderLandingPage from '../components/RenderLandingPage';
 
 function LandingContainer({ LoadingComponent }) {
   return (
