+++ src/components/login/components/index.js
@@ -1,0 +1,1 @@
+export { default as LoginPage } from '../pages/LoginContainer';
