--- src/components/pages/ExampleList/index.js
@@ -1,1 +1,0 @@
-export { default as ExampleListPage } from './ExampleListContainer';
