--- src/components/pages/Landing/index.js
@@ -1,1 +1,0 @@
-export { default as LandingPage } from './LandingContainer';
