+++ src/components/profileList/components/index.js
@@ -1,0 +1,1 @@
+export { default as ProfileListPage } from '../pages/ProfileListContainer';
