+++ src/components/profileList/pages/ProfileListContainer.js
@@ -1,23 +1,27 @@
 import React from 'react';
+import { useOktaAuth } from '@okta/okta-react';
 
-import { getExampleData } from '../../../api';
+import { getProfileData } from '../../../api';
 
 import { List } from '../../common';
-import RenderExampleListPage from './RenderExampleListPage';
+
+import RenderProfileListPage from '../components/RenderProfileListPage';
 
 // Here is an example of using our reusable List component to display some list data to the UI.
-const ExampleList = () => {
+const ProfileList = () => {
+  const { authState } = useOktaAuth();
+
   return (
     <List
       // Here we are passing our Axios request helper function as a callback.
-      getItemsData={getExampleData}
+      getItemsData={() => getProfileData(authState)}
       // Here we are passing in a component we want to show whilst waiting for our API request
       // to complete.
-      LoadingComponent={() => <div>Loading Items...</div>}
+      LoadingComponent={() => <div>Loading Profiles...</div>}
       // Here we are passing in a component that receives our new data and returns our JSX elements.
-      RenderItems={RenderExampleListPage}
+      RenderItems={RenderProfileListPage}
     />
   );
 };
 
-export default ExampleList;
+export default ProfileList;
